fix(variants): SJRA-1564 seq id ans task id params get together
